Mainly changes to the pump_driver ketnel module
Added some code to use available counters in the RaspberrPi ARM chip to
determine the width of pulses. This is used to avoid spurious pulses on
the load signal.
Added the ability to handle 'blank' seven-segment character to the
serial decoder code.
run.bash is written to set the paths properly and make it possible to
run perl apps with 'sudo' and use a perlbrew perl.